backend/pyproject.toml
[project]
name = "askfia-api"
version = "0.2.0"
description = "AskFIA - AI Conversational Interface for Forest Inventory Analysis - Part of the FIA Python Ecosystem"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Chris Mihiar", email = "chris@example.com" }
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyfia @ git+https://github.com/mihiarc/pyfia.git@ae0a3659300770674b01afff5a191bc321cdf0e8",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "redis>=5.0.0",
    "httpx>=0.28.0",
    "polars>=1.0.0",
    "boto3>=1.42.14",
    "aiofiles>=25.1.0",
    "pyjwt>=2.10.1",
    "bcrypt>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.urls]
Homepage = "https://github.com/mihiarc/askfia"
Repository = "https://github.com/mihiarc/askfia"
"Bug Tracker" = "https://github.com/mihiarc/askfia/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/askfia_api"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
